+++ CHANGELOG.md
@@ -125,8 +125,8 @@ Notes: web developers are advised to use [`~` (tilde range)](https://github.com/
    - The mode is suitable for providing chain-of-thought reasoning
    - Added visual indication of `creativeWorkStatus` property in `Message` entity:
       - `undefined` - no indicator is shown
-      - `'incomplete'` - show loader aside from the message
-      - `'published'` - show checkmark aside from the message
+      - `'Incomplete'` - show loader aside from the message
+      - `'Published'` - show checkmark aside from the message
    - Added a `position` property support for the `Message` entity
       - The `position` property takes precedence over existing sequence ID
 - Unified build info and injection of `<meta>` tag, in PR [#5571](https://github.com/microsoft/BotFramework-WebChat/pull/5571), by [@compulim](https://github.com/compulim)